+++ src/ui/identities/funding_common.rs
@@ -3,6 +3,11 @@ use eframe::epaint::{Color32, ColorImage};
 use egui::Vec2;
 use image::Luma;
 use qrcode::QrCode;
+use std::sync::{Arc, RwLock};
+
+use crate::model::wallet::Wallet;
+use dash_sdk::dashcore_rpc::dashcore::Address;
+use dash_sdk::dpp::dashcore::{OutPoint, TxOut};
 
 #[derive(Eq, PartialEq, Ord, PartialOrd, Copy, Clone)]
 pub enum WalletFundedScreenStep {
@@ -46,4 +51,44 @@ pub fn copy_to_clipboard(text: &str) -> Result<(), String> {
     clipboard
         .set_text(text.to_string())
         .map_err(|e| e.to_string())
+}
+
+pub fn capture_qr_funding_utxo_if_available(
+    step: &Arc<RwLock<WalletFundedScreenStep>>,
+    wallet: Option<&Arc<RwLock<Wallet>>>,
+    funding_address: Option<&Address>,
+) -> Option<(OutPoint, TxOut, Address)> {
+    if !matches!(
+        *step.read().expect("wallet funding step lock poisoned"),
+        WalletFundedScreenStep::WaitingOnFunds
+    ) {
+        return None;
+    }
+
+    let address = funding_address.cloned()?;
+
+    let wallet_arc = wallet?;
+
+    let candidate_utxo = {
+        let wallet = wallet_arc
+            .read()
+            .expect("wallet lock poisoned while checking funding UTXO");
+        wallet.utxos.get(&address).and_then(|utxos| {
+            utxos
+                .iter()
+                .filter(|(_, tx_out)| tx_out.value > 0)
+                .max_by_key(|(_, tx_out)| tx_out.value)
+                .map(|(outpoint, tx_out)| (*outpoint, tx_out.clone()))
+        })
+    };
+
+    if let Some((outpoint, tx_out)) = candidate_utxo {
+        let mut step = step
+            .write()
+            .expect("wallet funding step write lock poisoned");
+        *step = WalletFundedScreenStep::FundsReceived;
+        Some((outpoint, tx_out, address))
+    } else {
+        None
+    }
 }
