+++ src/ui/identities/top_up_identity_screen/by_wallet_qr_code.rs
@@ -1,7 +1,7 @@
 use crate::app::AppAction;
 use crate::backend_task::BackendTask;
 use crate::backend_task::identity::{IdentityTask, IdentityTopUpInfo, TopUpIdentityFundingMethod};
-use crate::ui::identities::funding_common::{copy_to_clipboard, generate_qr_code_image};
+use crate::ui::identities::funding_common::{self, copy_to_clipboard, generate_qr_code_image};
 use crate::ui::identities::top_up_identity_screen::{TopUpIdentityScreen, WalletFundedScreenStep};
 use dash_sdk::dashcore_rpc::RpcApi;
 use eframe::epaint::TextureHandle;
@@ -93,7 +93,13 @@ impl TopUpIdentityScreen {
 
     pub fn render_ui_by_wallet_qr_code(&mut self, ui: &mut Ui, step_number: u32) -> AppAction {
         // Update state when the QR funding address receives funds
-        self.capture_qr_funding_utxo_if_available();
+        if let Some(utxo) = funding_common::capture_qr_funding_utxo_if_available(
+            &self.step,
+            self.wallet.as_ref(),
+            self.funding_address.as_ref(),
+        ) {
+            self.funding_utxo = Some(utxo);
+        }
 
         // Extract the step from the RwLock to minimize borrow scope
         let step = *self.step.read().unwrap();
